% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domultiheatmap.func.R
\name{DoMultiBarHeatmap}
\alias{DoMultiBarHeatmap}
\title{DoMultiBarHeatmap}
\arguments{
\item{object}{Seurat object}

\item{features}{A vector of features to plot, defaults to VariableFeatures(object = object)}

\item{cells}{A vector of cells to plot}

\item{group.by}{A vector of variables to group cells by; defaults to cell identity classes}

\item{additional.group.by}{A second vector of variables to group cells by}

\item{group.bar}{Add a color bar showing group status for cells (boolean)}

\item{group.colors}{Colors to use for the color bar}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped); defaults to 2.5 if slot is 'scale.data', 6 otherwise}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{assay}{Assay to pull from}

\item{label}{Label the cell identies above the color bar}

\item{size}{Size of text above color bar}

\item{hjust}{Horizontal justification of text above color bar}

\item{angle}{Angle of text above color bar}

\item{raster}{If true, plot with geom_raster, else use geom_tile. 
geom_raster may look blurry on some viewing applications such as Preview 
due to how the raster is interpolated. Set this to FALSE if you are 
encountering that issue (note that plots may take longer to produce/render).}

\item{draw.lines}{Include white lines to separate the groups}

\item{lines.width}{Integer number to adjust the width of the separating 
white lines. Corresponds to the number of "cells" between each group.}

\item{group.bar.height}{Scale the height of the color bar}

\item{combine}{Combine plots into a single patchworked ggplot object.
If FALSE, return a list of ggplot objects}
}
\description{
Builds on Seurat's Doheatmap function code to produce a heatmap from a Seurat object with multiple annotation bars.
}
\examples{
data("pbmc_small")
DoMultiBarHeatmap(object = pbmc_small, group.by="cluster", additional.group.by="cell_type")

}
